Update description for Google Udacity course
Hey Amberley, I was just checking out the site after discovering it on [the State of JS report](https://2020.stateofjs.com/en-US/features/syntax/).

I updated the description for the Udacity course that Alice and I worked on. We didn't work with AT&T on it—maybe Udacity was showing AT&T's logo somewhere on the page to promote something else?